remove unneeded tearDown method to avoid deadlock
calling saveMainContextAndWait in tearDown was causing a deadlock on Travis (that was reproducable when limited my machine to 1 CPU).  The resetCoreData call in setUp for the following test was deadlocking when trying to remove the peristent store.  I suspect due to the root context still in the process of saving changes.